refactor(unity-exec): extract plan token module from phase executor
Why:
- OperationPhaseExecutor に plan token 関連責務が集中しており、実装凝集が低く変更影響が広がりやすかったため

What Changed:
- plan token 契約/環境/指紋計算/coordinator を Editor/Execution/PlanToken 配下へ分割した
- OperationPhaseExecutor を実行オーケストレーション専任へ整理し、PlanToken namespace 参照に変更した
- OperationPhaseExecutorTests に PlanToken namespace の using を追加した

Impact:
- 実行制御と token 基盤の責務境界が明確になり、仕様変更時の影響範囲を局所化できる
