+++ src/Ucli.Unity/Assets/MackySoft/MackySoft.Ucli.Unity/Editor/Execution/PlanToken.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 34f3003ec9eef482fa6eec0cb2d8b3cf
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
