+++ src/Ucli.Unity/Assets/MackySoft/MackySoft.Ucli.Unity/Editor/Execution/PlanToken/UnityProjectFingerprintCalculatorCompat.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0be98f029119640b89b868bba3fdcddd
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
